show correct software table data
