+++ frontend/pages/SoftwarePage/SoftwareTitles/SoftwareTable/_styles.scss
@@ -1,0 +1,32 @@
+.software-table {
+  &__count {
+    display: flex;
+    gap: 12px;
+  }
+
+  &__vuln_dropdown {
+    .Select-menu-outer {
+      width: 250px;
+      max-height: 310px;
+
+      .Select-menu {
+        max-height: none;
+      }
+    }
+
+    .Select-value {
+      padding-left: $pad-medium;
+      padding-right: $pad-medium;
+    }
+
+    .dropdown__custom-value-label {
+      width: 155px; // Override 105px for longer text options
+    }
+  }
+
+  &__filter-controls {
+    display: flex;
+    align-items: center;
+    gap: $pad-medium;
+  }
+}
