+++ frontend/pages/SoftwarePage/SoftwareTitles/SoftwareTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SoftwareTable";
