--- frontend/pages/SoftwarePage/SoftwareVersions/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SoftwareVersions";
